Update README.md for Confluent Parallel Consumer application
This commit elaborates on the Confluent Parallel Consumer application details in the README.md. It expands on the level of parallelism achievable beyond the partition count. It also provides instructions on creating a new instance of `ParallelStreamProcessor` for parallel record consuming and introduces available performance tests for the parallel consumption. This update strengthens the guide for users exploring this library.