Conversion of column-diff, concatentation and parallel consumer (#6)
* column diff conversion

* Conversion of concatenation tutorial

* First commit for parallel consumer conversion

* Delete gradle.properties in cogrouping-streams/kstreams

This commit removes the gradle.properties file from the cogrouping-streams/kstreams directory. This is part of the initial steps in converting to a parallel consumer setup.

* column diff conversion

* Conversion of concatenation tutorial

* First commit for parallel consumer conversion

* Delete gradle.properties in cogrouping-streams/kstreams

This commit removes the gradle.properties file from the cogrouping-streams/kstreams directory. This is part of the initial steps in converting to a parallel consumer setup.

* Update README.md for Confluent Parallel Consumer application

This commit elaborates on the Confluent Parallel Consumer application details in the README.md. It expands on the level of parallelism achievable beyond the partition count. It also provides instructions on creating a new instance of `ParallelStreamProcessor` for parallel record consuming and introduces available performance tests for the parallel consumption. This update strengthens the guide for users exploring this library.

* Apply suggestions from code review

Co-authored-by: Dave Troiano <dtroiano@confluent.io>

---------

Co-authored-by: Dave Troiano <dtroiano@confluent.io>