Update count-messages (ksqlDB), first-kafka-streams-application (Kafka Streams), deserialization-errors (ksqlDB) (#25)
* Simplification of the counting topics and first Kafka Streams application.

* Add README for handling ksqlDB deserialization errors

A detailed guide on identifying and managing deserialization errors in ksqlDB has been added. This includes a walkthrough on setting up a program that monitors a stream of sensors with a focus on detecting and resolving deserialization errors that may prevent events from being written into a stream or table.

* converted the ksqlDB deserialization tutorial

* Update indentation in SQL statement

The adjustment provides better readability in the SQL code on the ksqlDB deserialization tutorial file. The change only affects the aesthetics and does not alter the functionality or output of the code.

* Fix markdown hyperlink and refine README.md content

This commit corrects a hyperlink syntax in the deserialization-errors tutorial and improves wording in several README.md files. These updates enhance clarity for users, providing a neater presentation and more precise language for superior tutorial readability.