Update Flink SQL testing approach (#40)
* Update test cases for FlinkSqlTumblingWindow

The changes modify test cases for the Flink SQL Tumbling Window by altering the test resources, and updating the Java test code. The modifications are carried towards obtaining and comparing tabular results from the SQL execution, which aims to make test cases more comprehensive and reliable.

* Remove unused imports in FlinkSqlTumblingWindowTest

This commit removes unused import statements in the FlinkSqlTumblingWindowTest file, making the code cleaner and easier to read. The modifications include the removal of the EnvironmentSettings, Table, and TableEnvironment classes from org.apache.flink.table.api, along with several assertion methods and expressions.

* Refactor test utils into a separate class

Moved utilities for getting row objects from TableResults and parsing datetime strings into a separate TestUtils class. This helps in reducing clutter in FlinkSqlTumblingWindowTest and allows reusable utility functions to be used in other tests as well. The changes improve readability and maintainability of the test code.

* Remove expected-ratings-by-6hr-window test resource file

The test resources file "expected-ratings-by-6hr-window.txt" has been deleted. This file was previously used for testing ratings and their respective window times, but it is no longer necessary for the project.