Update Flink SQL tests and documentation for Top-N queries
The Flink SQL tests and README documentation were updated to reflect a change in the number of top movies we are retrieving based on genre, changing it from top 5 to top 3. Also, the value format for the movie_views table was changed from 'avro-confluent' to 'json'. The structure and the content of the SQL queries were also modified accordingly.