Update wording in README.md
The README document for the top-N FlinkSQL has been updated. The changes improve the explanation about what service or product the user might want to monitor. Now it clearly states that the person might need to monitor some service or product, instead of assuming they have some service or product.