feat: Add Top-N Tutorial (#43)
* Add SQL scripts and unit test for Top-N Flink tutorial

This commit introduces new SQL scripts, a unit test, and related files for a tutorial about how to create a top-N query with Flink SQL.

* Refactor SQL tests to analyze movie views instead of sales

Renamed and updated SQL queries in the test resources to focus on analyzing movie views per genre. These changes allow the computation of a top-N list of movies per genre, based on view counts, instead of the minimum and maximum sales from previous years. The README was also updated to match these changes. Updated dependencies in build.gradle file.

* Update Flink SQL tests and documentation for Top-N queries

The Flink SQL tests and README documentation were updated to reflect a change in the number of top movies we are retrieving based on genre, changing it from top 5 to top 3. Also, the value format for the movie_views table was changed from 'avro-confluent' to 'json'. The structure and the content of the SQL queries were also modified accordingly.

* Apply suggestions from code review

Co-authored-by: Dave Troiano <dtroiano@confluent.io>

* Update wording in README.md

The README document for the top-N FlinkSQL has been updated. The changes improve the explanation about what service or product the user might want to monitor. Now it clearly states that the person might need to monitor some service or product, instead of assuming they have some service or product.

---------

Co-authored-by: Dave Troiano <dtroiano@confluent.io>