Add KStreams application to join movie and rating streams
The commit introduces a Kafka Streams (KStreams) application named `JoinStreamToStream` which joins two streams of data; movies and ratings. This application uses `KStream.join` operation to combine each rating with the corresponding movie. This commit also includes the testing of this joining operation. Movie and rating data models are also added along with a `MovieRatingJoiner` to specify how to construct the joined value of both streams.