feat: Add example of reordering the events of a Kafka Streams application (#53)
* Add stream reordering functionality for Kafka Streams

Added functionality to reorder Kafka Streams by the timestamp embedded in the message payload. Created new Java classes ReorderStreams, ReorderProcessorSupplier and Event, and respective unit tests to support the functionality. Updated the project structure and build.gradle as needed.

* Refactor test method and update dependencies in ReorderStreams module

Imported java.time.Instant and refactored the test method in the ReorderProcessorTest class for better readability. Also updated the version of Kafka streams and clients dependencies from 3.6.0 to 3.7.0 in the build.gradle files of both the ReorderStreams and common modules.

* Update README and ReorderingProcessorSupplier for reordering out-of-order events

The README for the reordering-streams/kstreams module has been significantly improved to provide more detailed instructions on how to reorder out-of-order events in Kafka Streams. Additionally, the ReorderingProcessorSupplier.java file has been slightly edited to improve its readability.

* Update reordering streams code and documentation

The method name `punctuate` was updated to `forwardOrderedByEventTime` to more accurately reflect its function. The variable `grace` was also renamed to `reorderWindow` for more clarity. Both code and README file were updated with a detailed explanation of the reordering strategy for events in the stream; including aspects such as restoring the original key, limitations to a single partition, and considerations for choosing the window size for reordering.

* Remove unnecessary comments and redundant lines

The unnecessary comments from several files, specifically Event.java and ReorderingProcessorSupplier.java, are removed. Also, redundant lines in the markdown file README.md are deleted for a cleaner and more concise document.

* Apply suggestions from code review

Co-authored-by: Dave Troiano <dtroiano@confluent.io>

---------

Co-authored-by: Dave Troiano <dtroiano@confluent.io>